Add files via upload
A factor is a number that divides another number, leaving no remainder.
Factorial, in general, is represented as n!, which is equal to n* (n-1)* (n-2)* (n-3)*….*1, where n can be any finite number. In Python, Factorial can be achieved by a loop function, defining a value for n or passing an argument to create a value for n or creating a prompt to get the user’s desired input.